all parity tests pass
