+++ pax/conf/experiment/hyper.yaml
@@ -9,7 +9,7 @@ env_id: ipd
 game: ipd 
 env_type: infinite
 env_discount: 0.99
-payoff:
+payoff: [[-1, -1], [-3, 0], [0, -3], [-2, -2]]
 
 # Training hyperparameters
 num_envs: 50
