+++ pax/conf/experiment/naive.yaml
@@ -14,8 +14,8 @@ payoff: [[-1, -1], [-3, 0], [0, -3], [-2, -2]]
 # Training hyperparameters
 num_envs: 100
 num_steps: 150 # number of steps per episode
-total_timesteps: 400_000
-eval_every: 100_000 # timesteps
+total_timesteps: 4_000_000
+eval_every: 1_000_000 # timesteps
 
 # Useful information
 # num_episodes = total_timesteps / num_steps
